+++ test/TestMultipleAlgoExecutions.cpp
@@ -3,6 +3,8 @@
 
 #include "kompute/Kompute.hpp"
 
+#include "kompute_test/Shader.hpp"
+
 TEST(TestMultipleAlgoExecutions, TestEndToEndFunctionality)
 {
 
@@ -51,7 +53,7 @@ TEST(TestMultipleAlgoExecutions, TestEndToEndFunctionality)
     kp::Constants pushConstsB({ 3.0 });
 
     auto algorithm = mgr.algorithm(params,
-                                   kp::Shader::compileSource(shader),
+                                   kp_test_utils::Shader::compileSource(shader),
                                    workgroup,
                                    specConsts,
                                    pushConstsA);
@@ -89,7 +91,7 @@ TEST(TestMultipleAlgoExecutions, SingleSequenceRecord)
           pa[index] = pa[index] + 1;
       })");
 
-    std::vector<uint32_t> spirv = kp::Shader::compileSource(shader);
+    std::vector<uint32_t> spirv = kp_test_utils::Shader::compileSource(shader);
 
     {
         // A sharedMemoryBarrier is required as the shader is not thread-safe:w
@@ -130,7 +132,7 @@ TEST(TestMultipleAlgoExecutions, MultipleCmdBufRecords)
           pa[index] = pa[index] + 1;
       })");
 
-    std::vector<uint32_t> spirv = kp::Shader::compileSource(shader);
+    std::vector<uint32_t> spirv = kp_test_utils::Shader::compileSource(shader);
 
     std::shared_ptr<kp::Algorithm> algorithm =
       mgr.algorithm({ tensorA }, spirv);
@@ -166,7 +168,7 @@ TEST(TestMultipleAlgoExecutions, MultipleSequences)
           pa[index] = pa[index] + 1;
       })");
 
-    std::vector<uint32_t> spirv = kp::Shader::compileSource(shader);
+    std::vector<uint32_t> spirv = kp_test_utils::Shader::compileSource(shader);
 
     std::shared_ptr<kp::Algorithm> algorithm =
       mgr.algorithm({ tensorA }, spirv);
@@ -201,7 +203,7 @@ TEST(TestMultipleAlgoExecutions, SingleRecordMultipleEval)
           pa[index] = pa[index] + 1;
       })");
 
-    std::vector<uint32_t> spirv = kp::Shader::compileSource(shader);
+    std::vector<uint32_t> spirv = kp_test_utils::Shader::compileSource(shader);
 
     std::shared_ptr<kp::Algorithm> algorithm =
       mgr.algorithm({ tensorA }, spirv);